userspace: split runtime policy-ID helpers out of policies.go
Move the runtime policy-ID / stable-identity concern verbatim from
policies.go into a new policies_ids.go: RuntimePolicyIDs,
PolicyIDsByStableKey, PoliciesByStableKey, stablePolicyRuleID,
StablePolicyRuleID, RuntimePolicyIndex, userspacePolicyRuleExpansionCount.
These derive the span-accumulated runtime/RT_FLOW policy IDs and the stable
"<from>-><to>/<name>" identities the display and #4234 deletion-clear paths
key off (#3056/#3063/#3143/#3336).

The policyRuleSlot type + policyID method and walkPolicyRuleSlots (the SSOT
namespace walk) stay in the residual policies.go: they are shared between
this ID concern and the lowering path, so they are the shared core.

Pure code-motion, no logic change: byte-identical slices of the
origin/master policies.go. Refs #4421 (codex-173 #4).
